Small changes.
Added another way to detect steam path (Should be somewhat cross-platform too! Requires Steam to be running)

Re-arranged some imports.
youtube-dl will be imported late, when the user presses the downloader button. The startup time is too high because of it, so we're loading it somewhat lazily.

Change the exception hook/wrapper to be a bit neater.

More logging stuff

Added youtube-dl to the Licenses dialog.